ap work & JBan changes

+++ src/main/java/me/pugabyte/bncore/features/commands/staff/UnbanCommand.java
@@ -1,12 +1,14 @@
 package me.pugabyte.bncore.features.commands.staff;
 
 import me.pugabyte.bncore.framework.commands.models.CustomCommand;
+import me.pugabyte.bncore.framework.commands.models.annotations.Arg;
 import me.pugabyte.bncore.framework.commands.models.annotations.Fallback;
 import me.pugabyte.bncore.framework.commands.models.annotations.Path;
 import me.pugabyte.bncore.framework.commands.models.annotations.Permission;
 import me.pugabyte.bncore.framework.commands.models.events.CommandEvent;
 import me.pugabyte.bncore.models.delayedban.DelayedBan;
 import me.pugabyte.bncore.models.delayedban.DelayedBanService;
+import org.bukkit.Bukkit;
 import org.bukkit.OfflinePlayer;
 
 @Fallback("litebans")
@@ -25,12 +27,15 @@ public class UnbanCommand extends CustomCommand {
 	}
 
 	@Path("<player>")
-	public void unban(OfflinePlayer player) {
-		DelayedBanService delayedBanService = new DelayedBanService();
-		if (delayedBanService.hasQueuedBan(player)) {
-			DelayedBan delayedBan = delayedBanService.get(player.getUniqueId());
-			delayedBanService.deleteSync(delayedBan);
-			send(DelayedBanCommand.PREFIX + player.getName() + "'s delayed ban removed");
+	public void unban(@Arg(tabCompleter = OfflinePlayer.class) String playerName) {
+		OfflinePlayer player = Bukkit.getPlayer(playerName);
+		if (player != null && player.hasPlayedBefore()) {
+			DelayedBanService delayedBanService = new DelayedBanService();
+			if (delayedBanService.hasQueuedBan(player)) {
+				DelayedBan delayedBan = delayedBanService.get(player.getUniqueId());
+				delayedBanService.deleteSync(delayedBan);
+				send(DelayedBanCommand.PREFIX + player.getName() + "'s delayed ban removed");
+			}
 		}
 
 		fallback();
